refactor(session)!: 移除固定会话（Const）功能与 AI 标题生成
