+++ .gitignore
@@ -42,9 +42,6 @@ Thumbs.db
 # Claude Code
 .claude/
 
-# 用户固定会话（Const Session）持久化数据
-api/data/const-sessions/
-
 # Project specific
 botpy.log
 *.dll
