+++ README.md
@@ -89,15 +89,6 @@ start.bat
 
 > **提示**：如果提示"暂无已配置的 LLM 提供商"，请先前往侧栏 **设置 → 模型** 添加 API Key。
 
-## 固定会话
-
-固定会话（Const Session）是可以**永久保存**的聊天会话，不会因服务重启或 TTL 过期而丢失。
-
-- **固定会话**：在侧边栏会话列表右键任一普通会话 → **固定会话**，输入名称即可保存。点击按钮可自动生成名称。
-- **解绑固定**：右键已固定的会话 → **取消固定**
-- 固定会话的数据存储在 `api/data/const_sessions.yaml`，服务启动时自动加载
-- 适合保存常用的工具性会话、特定角色的对话场景
-
 ## 设置性格与用户偏好
 
 在页面栏 **设置 → 人设 / 用户** 中可直接在线编辑，自定义 Sonetto 的性格和它对你的了解：
