+++ api/utils/__init__.py
@@ -1,9 +1,11 @@
 """工具模块。"""
 
 from .logger import get_logger, set_trace_id, get_trace_id
+from .messages import flatten_content
 
 __all__ = [
     "get_logger",
     "set_trace_id",
     "get_trace_id",
+    "flatten_content",
 ]
