+++ web/src/components/SessionSidebar.vue
@@ -6,103 +6,46 @@
     </div>
     <div class="session-list">
 
-      <!-- ── 已保存（固定会话）── -->
-      <div class="const-section">
-        <div class="section-label">已保存</div>
-        <div v-if="constSessions.length === 0" class="section-hint">
-          无已保存的会话。
-          <br>
-          右键点击临时会话来固定保存
+      <!-- ── 会话列表 ── -->
+      <div
+        v-for="(s, index) in sessions"
+        :key="s.session_id"
+        class="session-item"
+        :class="{ active: s.session_id === activeId }"
+        @click="$emit('switch', s.session_id)"
+        @mouseenter="onSessionMouseEnter($event, s)"
+        @mouseleave="onSessionMouseLeave"
+      >
+        <div class="session-item-main">
+          <span class="session-id">
+            Session #{{ index + 1 }}
+            <span v-if="s.is_subagent" class="sub-badge" title="子 Agent 会话（只读）">sub</span>
+          </span>
+          <span class="session-count">{{ formatRelativeTime(s.last_active ?? s.created_at) }} · {{ s.message_count }} 条消息</span>
         </div>
-        <div
-          v-for="(s, ci) in constSessions"
-          :key="s.session_id"
-          class="session-item is-const"
-          :class="{ active: s.session_id === activeId }"
-          @click="$emit('switch', s.session_id)"
-          @contextmenu.prevent="onSessionContextMenu($event, s)"
-          @mouseenter="onSessionMouseEnter($event, s)"
-          @mouseleave="onSessionMouseLeave"
-        >
-          <div class="session-item-main">
-            <span class="session-id">
-              <Icon name="pin" :size="12" style="margin-right: 3px; flex-shrink: 0;" />
-              <span class="const-name-text">{{ s.const_name || '未命名' }}</span>
-              <span v-if="s.is_subagent" class="sub-badge" title="子 Agent 会话（只读）">sub</span>
-            </span>
-            <span class="session-count">{{ formatRelativeTime(s.last_active ?? s.created_at) }} · {{ s.message_count }} 条消息</span>
-          </div>
-          <div class="session-item-right">
-            <span
-              v-if="(sessionStatuses ?? {})[s.session_id]?.isAwaitingUser"
-              class="status-dot awaiting-user"
-              title="等待用户输入"
-            />
-            <span
-              v-else-if="(sessionStatuses ?? {})[s.session_id]?.isStreaming"
-              class="status-dot streaming"
-              title="Agent 运行中"
-            />
-            <span
-              v-else-if="(sessionStatuses ?? {})[s.session_id]?.connected"
-              class="status-dot connected"
-              title="已连接"
-            />
-            <button
-              class="btn-delete"
-              @click.stop="$emit('delete', s.session_id)"
-              title="删除会话"
-            >
-              &times;
-            </button>
-          </div>
-        </div>
-      </div>
-
-      <!-- ── 临时会话 ── -->
-      <div class="temp-section">
-        <div class="section-label">临时会话</div>
-        <div
-          v-for="(s, index) in tempSessions"
-          :key="s.session_id"
-          class="session-item"
-          :class="{ active: s.session_id === activeId }"
-          @click="$emit('switch', s.session_id)"
-          @contextmenu.prevent="onSessionContextMenu($event, s)"
-          @mouseenter="onSessionMouseEnter($event, s)"
-          @mouseleave="onSessionMouseLeave"
-        >
-          <div class="session-item-main">
-            <span class="session-id">
-              Session #{{ getSessionDisplayIndex(s) }}
-              <span v-if="s.is_subagent" class="sub-badge" title="子 Agent 会话（只读）">sub</span>
-            </span>
-            <span class="session-count">{{ formatRelativeTime(s.last_active ?? s.created_at) }} · {{ s.message_count }} 条消息</span>
-          </div>
-          <div class="session-item-right">
-            <span
-              v-if="(sessionStatuses ?? {})[s.session_id]?.isAwaitingUser"
-              class="status-dot awaiting-user"
-              title="等待用户输入"
-            />
-            <span
-              v-else-if="(sessionStatuses ?? {})[s.session_id]?.isStreaming"
-              class="status-dot streaming"
-              title="Agent 运行中"
-            />
-            <span
-              v-else-if="(sessionStatuses ?? {})[s.session_id]?.connected"
-              class="status-dot connected"
-              title="已连接"
-            />
-            <button
-              class="btn-delete"
-              @click.stop="$emit('delete', s.session_id)"
-              title="删除会话"
-            >
-              &times;
-            </button>
-          </div>
+        <div class="session-item-right">
+          <span
+            v-if="(sessionStatuses ?? {})[s.session_id]?.isAwaitingUser"
+            class="status-dot awaiting-user"
+            title="等待用户输入"
+          />
+          <span
+            v-else-if="(sessionStatuses ?? {})[s.session_id]?.isStreaming"
+            class="status-dot streaming"
+            title="Agent 运行中"
+          />
+          <span
+            v-else-if="(sessionStatuses ?? {})[s.session_id]?.connected"
+            class="status-dot connected"
+            title="已连接"
+          />
+          <button
+            class="btn-delete"
+            @click.stop="$emit('delete', s.session_id)"
+            title="删除会话"
+          >
+            &times;
+          </button>
         </div>
       </div>
 
@@ -138,68 +81,12 @@
       </div>
     </Transition>
 
-    <!-- ── 右键菜单 ── -->
-    <ContextMenu
-      :position="ctxMenuPos"
-      :items="ctxMenuItems"
-      :visible="ctxMenuVisible"
-      @select="handleContextMenuSelect"
-      @close="closeContextMenu"
-    />
-
-    <!-- ── 固定会话卡片 ── -->
-    <Teleport to="body">
-      <Transition name="constify-pop">
-        <div
-          v-if="constifyTarget"
-          ref="constifyCardRef"
-          class="constify-card"
-          :style="constifyCardStyle"
-          @click.stop
-        >
-          <div class="constify-card-title">固定会话</div>
-          <div v-if="constifyError" class="constify-error">{{ constifyError }}</div>
-          <div class="constify-input-row">
-            <input
-              ref="constifyInputRef"
-              v-model="constifyName"
-              class="constify-input"
-              type="text"
-              placeholder="输入会话名称..."
-              maxlength="50"
-              @keydown.enter="confirmConstify"
-              @keydown.esc="cancelConstify"
-            />
-            <button
-              class="constify-gen-btn"
-              :class="{ loading: generating }"
-              title="AI 生成标题"
-              :disabled="generating"
-              @click="generateTitle"
-            >
-              <Icon name="sparkles" :size="16" />
-            </button>
-          </div>
-          <div class="constify-actions">
-            <button class="constify-btn cancel" @click="cancelConstify">取消</button>
-            <button
-              class="constify-btn confirm"
-              :disabled="!constifyName.trim()"
-              @click="confirmConstify"
-            >确定</button>
-          </div>
-        </div>
-      </Transition>
-    </Teleport>
   </div>
 </template>
 
 <script setup lang="ts">
-import ContextMenu from '@/components/ContextMenu.vue';
-import Icon from '@/components/Icon.vue';
-import { generateSessionTitle } from '@/composables/useSession';
 import type { SessionInfo } from '@/types';
-import { computed, nextTick, ref, watch } from 'vue';
+import { computed, nextTick, ref } from 'vue';
 
 const props = defineProps<{
   sessions: SessionInfo[]
@@ -208,27 +95,11 @@ const props = defineProps<{
   collapsed?: boolean
 }>()
 
-const emit = defineEmits<{
+defineEmits<{
   create: []
   switch: [id: string]
   delete: [id: string]
-  constify: [id: string, name: string]
-  unconstify: [id: string]
 }>()
-
-// ── 分区计算 ──────────────────────────────────────────────────
-
-function getSessionDisplayIndex(s: SessionInfo): number {
-  return tempSessions.value.length - tempSessions.value.findIndex(x => x.session_id === s.session_id)
-}
-
-const constSessions = computed(() =>
-  props.sessions.filter(s => s.is_const)
-)
-
-const tempSessions = computed(() =>
-  props.sessions.filter(s => !s.is_const)
-)
 
 // ── Hover card ────────────────────────────────────────────────
 
@@ -307,170 +178,6 @@ const cardStyle = computed(() => {
     pointerEvents: 'none' as const,
   }
 })
-
-// ── 右键菜单 ──────────────────────────────────────────────────
-
-const ctxMenuVisible = ref(false)
-const ctxMenuPos = ref({ x: 0, y: 0 })
-const ctxMenuSession = ref<SessionInfo | null>(null)
-
-const ctxMenuItems = computed(() => {
-  const s = ctxMenuSession.value
-  if (!s) return []
-  if (s.is_const) {
-    return [
-      { label: '固定会话…', action: 'constify', icon: 'pin' },
-      { label: '取消固定', action: 'unconstify' },
-    ]
-  }
-  return [
-    { label: '固定会话…', action: 'constify', icon: 'pin' },
-  ]
-})
-
-// ── 固定会话卡片 ──────────────────────────────────────────────
-
-const constifyTarget = ref<SessionInfo | null>(null)
-const constifyName = ref('')
-const constifyCardRef = ref<HTMLElement | null>(null)
-const constifyInputRef = ref<HTMLInputElement | null>(null)
-const constifyCardTop = ref(0)
-const constifyCardLeft = ref(0)
-const generating = ref(false)
-const constifyError = ref('')
-
-/** 鼠标右键触发时的目标元素 rect，用于定位卡片 */
-let constifyAnchorRect: DOMRect | null = null
-
-const constifyCardStyle = computed(() => {
-  if (!constifyTarget.value) return { display: 'none' }
-  return {
-    position: 'fixed' as const,
-    top: `${constifyCardTop.value}px`,
-    left: `${constifyCardLeft.value}px`,
-    zIndex: 1001,
-  }
-})
-
-function adjustConstifyCardPosition() {
-  const cardEl = constifyCardRef.value
-  if (!cardEl || !constifyTarget.value) return
-  const cardWidth = cardEl.offsetWidth
-  const cardHeight = cardEl.offsetHeight
-  const vw = window.innerWidth
-  const vh = window.innerHeight
-  const margin = 8
-
-  if (constifyCardLeft.value + cardWidth > vw - margin) {
-    const leftPos = (constifyAnchorRect?.left ?? constifyCardLeft.value) - cardWidth - margin
-    constifyCardLeft.value = leftPos >= margin ? leftPos : Math.max(margin, vw - cardWidth - margin)
-  }
-  if (constifyCardTop.value + cardHeight > vh - margin) {
-    constifyCardTop.value = Math.max(margin, vh - cardHeight - margin)
-  }
-  // Ensure card stays on the right side of the anchor when there's room
-  if (constifyAnchorRect && constifyCardLeft.value < constifyAnchorRect.right) {
-    constifyCardLeft.value = constifyAnchorRect.right + margin
-  }
-}
-
-function showConstifyCard(session: SessionInfo) {
-  const rect = constifyAnchorRect
-  if (rect) {
-    constifyCardTop.value = rect.top
-    constifyCardLeft.value = rect.right + 8
-  }
-  constifyTarget.value = session
-  constifyName.value = session.const_name || ''
-  constifyError.value = ''
-  closeContextMenu()
-  nextTick(() => {
-    adjustConstifyCardPosition()
-    constifyInputRef.value?.focus()
-    constifyInputRef.value?.select()
-  })
-}
-
-async function generateTitle() {
-  const session = constifyTarget.value
-  if (!session || generating.value) return
-  generating.value = true
-  try {
-    const title = await generateSessionTitle(session.session_id)
-    constifyName.value = title
-    nextTick(() => {
-      constifyInputRef.value?.focus()
-      constifyInputRef.value?.select()
-    })
-  } catch (e) {
-    console.error('[constify] 标题生成失败:', e)
-    constifyError.value = e instanceof Error ? e.message : '标题生成失败'
-  } finally {
-    generating.value = false
-  }
-}
-
-function confirmConstify() {
-  if (!constifyTarget.value || !constifyName.value.trim()) return
-  emit('constify', constifyTarget.value.session_id, constifyName.value.trim())
-  constifyTarget.value = null
-  constifyName.value = ''
-}
-
-function cancelConstify() {
-  constifyTarget.value = null
-  constifyName.value = ''
-  constifyError.value = ''
-}
-
-// ── 悬浮详情卡片 ──────────────────────────────────────────────
-
-/** 注册一个全局点击监听，在 constify 卡片打开时点击外部关闭 */
-watch(constifyTarget, (val) => {
-  if (val) {
-    const handler = (e: MouseEvent) => {
-      const card = constifyCardRef.value
-      if (card && !card.contains(e.target as Node)) {
-        cancelConstify()
-      }
-    }
-    // delay registration so the current click doesn't immediately close
-    nextTick(() => document.addEventListener('click', handler, { once: true }))
-  }
-})
-
-function onSessionContextMenu(event: MouseEvent, session: SessionInfo) {
-  // 关闭 hover card
-  if (hoverLeaveTimer !== null) {
-    clearTimeout(hoverLeaveTimer)
-    hoverLeaveTimer = null
-  }
-  hoveredSession.value = null
-
-  // 保存锚点元素 rect，供 constify 卡片定位
-  constifyAnchorRect = (event.currentTarget as HTMLElement).getBoundingClientRect()
-
-  ctxMenuSession.value = session
-  ctxMenuPos.value = { x: event.clientX, y: event.clientY }
-  ctxMenuVisible.value = true
-}
-
-function handleContextMenuSelect(action: string) {
-  const s = ctxMenuSession.value
-  if (!s) return
-  if (action === 'constify') {
-    showConstifyCard(s)
-    return  // 不关闭菜单，showConstifyCard 内部会关闭
-  } else if (action === 'unconstify') {
-    emit('unconstify', s.session_id)
-  }
-  closeContextMenu()
-}
-
-function closeContextMenu() {
-  ctxMenuVisible.value = false
-  ctxMenuSession.value = null
-}
 </script>
 
 <style scoped>
@@ -523,23 +230,6 @@ function closeContextMenu() {
   padding: 0 6px 8px;
 }
 
-/* ── Section label ── */
-.section-label {
-  font-size: 10px;
-  font-weight: 600;
-  color: var(--text-tertiary, #9ca3af);
-  text-transform: uppercase;
-  letter-spacing: 0.4px;
-  padding: 4px 12px 2px;
-}
-
-.section-hint {
-  font-size: 11px;
-  color: var(--text-tertiary, #9ca3af);
-  padding: 6px 12px 8px;
-  line-height: 1.4;
-}
-
 /* ── Session items ── */
 .session-item {
   display: flex;
@@ -562,15 +252,6 @@ function closeContextMenu() {
   box-shadow: var(--shadow);
 }
 
-/* Const 会话外观 — pin 图标已提供视觉区分，无需侧边条 */
-.session-item.is-const {
-  background: color-mix(in srgb, var(--accent) 4%, transparent);
-}
-/* active const sessions should have a clean white bg */
-.session-item.is-const.active {
-  background: var(--bg-card);
-}
-
 .session-item-main {
   display: flex;
   flex-direction: column;
@@ -587,11 +268,6 @@ function closeContextMenu() {
   display: flex;
   align-items: center;
   gap: 2px;
-}
-.const-name-text {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
 }
 .sub-badge {
   display: inline-block;
@@ -756,165 +432,5 @@ function closeContextMenu() {
   overflow: hidden;
   padding: 0;
   margin: 0;
-}
-/* collapsed 时隐藏分区 label 和提示 */
-.session-sidebar.collapsed .section-label,
-.session-sidebar.collapsed .section-hint {
-  max-height: 0;
-  opacity: 0;
-  overflow: hidden;
-  padding: 0;
-  margin: 0;
-}
-/* collapsed 时隐藏 const 左侧边框 */
-.session-sidebar.collapsed .session-item.is-const {
-  border-left: none;
-}
-
-/* ── 固定会话卡片 ── */
-.constify-card {
-  min-width: 240px;
-  padding: 16px;
-  background: var(--bg-card);
-  border: 1px solid var(--border);
-  border-radius: 10px;
-  box-shadow: var(--shadow-xl);
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  pointer-events: auto;
-}
-
-.constify-card-title {
-  font-size: 14px;
-  font-weight: 600;
-  color: var(--text-primary);
-}
-
-.constify-error {
-  font-size: 12px;
-  color: #ef4444;
-  background: rgba(239, 68, 68, 0.08);
-  padding: 6px 10px;
-  border-radius: 6px;
-  line-height: 1.4;
-  word-break: break-all;
-}
-
-.constify-input-row {
-  display: flex;
-  gap: 6px;
-  align-items: stretch;
-}
-
-.constify-input {
-  flex: 1;
-  padding: 8px 12px;
-  font-size: 13px;
-  font-family: inherit;
-  color: var(--text-primary);
-  background: var(--bg-primary);
-  border: 1px solid var(--border);
-  border-radius: 6px;
-  outline: none;
-  transition: border-color 0.15s;
-}
-
-.constify-input:focus {
-  border-color: var(--accent);
-}
-
-.constify-input::placeholder {
-  color: var(--text-tertiary);
-}
-
-.constify-gen-btn {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 34px;
-  border: 1px solid var(--border);
-  border-radius: 6px;
-  background: var(--bg-primary);
-  color: var(--text-secondary);
-  cursor: pointer;
-  transition: background 0.15s, color 0.15s, border-color 0.15s;
-  flex-shrink: 0;
-}
-
-.constify-gen-btn:hover:not(:disabled) {
-  background: color-mix(in srgb, var(--accent) 8%, transparent);
-  color: var(--accent);
-  border-color: var(--accent);
-}
-
-.constify-gen-btn:disabled {
-  opacity: 0.4;
-  cursor: not-allowed;
-}
-
-.constify-gen-btn.loading {
-  animation: gen-btn-pulse 0.8s ease-in-out infinite;
-}
-
-@keyframes gen-btn-pulse {
-  0%, 100% { opacity: 0.5; }
-  50% { opacity: 1; }
-}
-
-.constify-actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 8px;
-}
-
-.constify-btn {
-  padding: 6px 16px;
-  font-size: 13px;
-  font-family: inherit;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-  transition: background 0.15s, opacity 0.15s;
-}
-
-.constify-btn.cancel {
-  background: transparent;
-  color: var(--text-secondary);
-}
-
-.constify-btn.cancel:hover {
-  background: color-mix(in srgb, var(--accent) 8%, transparent);
-  color: var(--text-primary);
-}
-
-.constify-btn.confirm {
-  background: var(--accent);
-  color: #fff;
-}
-
-.constify-btn.confirm:hover:not(:disabled) {
-  opacity: 0.85;
-}
-
-.constify-btn.confirm:disabled {
-  opacity: 0.4;
-  cursor: not-allowed;
-}
-
-/* 卡片弹出动画 */
-.constify-pop-enter-active {
-  transition: opacity 0.12s ease-out, transform 0.12s ease-out;
-}
-.constify-pop-leave-active {
-  transition: opacity 0.1s ease-in, transform 0.1s ease-in;
-}
-.constify-pop-enter-from {
-  opacity: 0;
-  transform: translateX(-6px) scale(0.96);
-}
-.constify-pop-leave-to {
-  opacity: 0;
-  transform: translateX(-4px) scale(0.96);
 }
 </style>
